# commerceue-app/WebContent/swagger_v3.0.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"API Extensions for IBM Digital Commerce customizations.",
      "version":"3.0",
      "title":"IBM Digital Commerce API Extensions"
   },
   "basePath":"/commerceue/extension",
...
   "paths":{  
...
      "/search/productview/postByCategory":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search response after getting product details based on the category.",
            "description":"",
            "operationId":"postFindProductsByCategory",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/SearchUEResponse"
                  }
               }
            }
         }
      },
      "/search/productview/postById":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search response after getting product details based on the product ID.",
            "description":"",
            "operationId":"postFindProductsById",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/SearchUEResponse"
                  }
               }
            }
         }
      },
      "/search/productview/postByIds":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search response after getting product details based on the product IDs.",
            "description":"",
            "operationId":"postFindProductsByIds",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/SearchUEResponse"
                  }
               }
            }
         }
      },
      "/search/productview/postByPartNumber":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search response after getting product details based on the part number.",
            "description":"",
            "operationId":"postFindProductsByPartNumber",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/SearchUEResponse"
                  }
               }
            }
         }
      },
      "/search/productview/postByPartNumbers":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search response after getting product details based on the part numbers.",
            "description":"",
            "operationId":"postFindProductsByPartNumbers",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/SearchUEResponse"
                  }
               }
            }
         }
      },
      "/search/productview/postBySearchTerm":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search response after getting product details based on a search term. For example, update the short description of the given product.",
            "description":"",
            "operationId":"postFindProductsBySearchTerm",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/SearchUEResponse"
                  }
               }
            }
         }
      },
      "/search/productview/preByCategory":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search expression before getting product details based on the category.",
            "description":"",
            "operationId":"preFindProductsByCategory",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
      "/search/productview/preById":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search expression before getting product details based on the product ID.",
            "description":"",
            "operationId":"preFindProductsById",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
      "/search/productview/preByIds":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search expression before getting product details based on the product IDs.",
            "description":"",
            "operationId":"preFindProductsByIds",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
      "/search/productview/preByPartNumber":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search expression before getting product details based on the part number.",
            "description":"",
            "operationId":"preFindProductsByPartNumber",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
      "/search/productview/preByPartNumbers":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search expression before getting product details based on the part numbers.",
            "description":"",
            "operationId":"preFindProductsByPartNumbers",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
      "/search/productview/preBySearchTerm":{  
         "post":{  
            "tags":[  
               "search - product view"
            ],
            "summary":"Customize the search expression before getting product details based on a search term.",
            "description":"",
            "operationId":"preFindProductsBySearchTerm",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"ProductViewPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/SearchUERequest"
                  }
               }
            ],
            "responses":{  
               "default":{  
                  "description":"successful operation"
               }
            }
         }
      },
...
   },
   "definitions":{  
...
      "ExceptionData":{  
         "type":"object",
         "properties":{  
            "code":{  
               "type":"string"
            },
            "message":{  
               "type":"string"
            },
            "messageKey":{  
               "type":"string"
            },
            "correlationId":{  
               "type":"string"
            },
            "messageArguments":{  
               "type":"array",
               "items":{  
                  "type":"string"
               }
            }
         }
      },
...
      "Locale":{  
         "type":"object",
         "properties":{  
            "language":{  
               "type":"string"
            },
            "script":{  
               "type":"string"
            },
            "country":{  
               "type":"string"
            },
            "variant":{  
               "type":"string"
            },
            "extensionKeys":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "unicodeLocaleAttributes":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "unicodeLocaleKeys":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "iso3Language":{  
               "type":"string"
            },
            "iso3Country":{  
               "type":"string"
            },
            "displayLanguage":{  
               "type":"string"
            },
            "displayScript":{  
               "type":"string"
            },
            "displayCountry":{  
               "type":"string"
            },
            "displayVariant":{  
               "type":"string"
            },
            "displayName":{  
               "type":"string"
            }
         }
      },
...
      "SearchUEResponse":{  
         "type":"object",
         "properties":{  
            "errors":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ExceptionData"
               }
            },
            "content":{  
               "type":"array",
               "items":{  
                  "type":"object",
                  "additionalProperties":{  
                     "type":"object"
                  }
               }
            }
         }
      },
...
      "SearchUERequest":{  
         "type":"object",
         "properties":{  
            "contextData":{  
               "$ref":"#/definitions/ContextData"
            },
            "content":{  
               "type":"array",
               "items":{  
                  "type":"object",
                  "additionalProperties":{  
                     "type":"object"
                  }
               }
            }
         }
      },
...
      "ContextData":{  
         "type":"object",
         "properties":{  
            "baseContext":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "accessControlContext":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "locale":{  
               "$ref":"#/definitions/Locale"
            }
         }
      },
...
   }
}